+++ ChangeLog.md
@@ -73,7 +73,7 @@
    * In order to make this compatible with Metronome Only mode, all pokemon that evolve due to knowing a move (e.g. Bonsly) now require Metronome regardless of whether evolution methods are randomized
    * Fixing impossible evos occurs after movesets are randomized, and selects a new move from the randomzied moveset to replace the current one used when evolving while knowing a move regardless of whether evolution methods are randomized
    * Evolutions like Nincada and Karrablast are excluded due to lack of support in game for randomization of these methods
- * Starter Pokemon selection has been condensed to use a slider to select whether the starters have at least 1, 2, or any amount of evolutions. 
+ * Starter Pokemon selection has been condensed to use a slider to select whether the starters have at least 1, 2, or any amount of evolutions
    * Added a checkbox to make starters have that exact count. 
    * This removes the ability to have 1 or 2 evo evolutions, but can be achieved by selecting "Limit Evolutions to Three Stages" when randomizing evolutions and selecting option 1 on the slider.
  * Fix Starters No Split to only activate when Random Starters is selected
@@ -81,3 +81,4 @@
  * Trainer pokemon held items can be randomized for Generation 3 or higher. Items are restricted to battle tower viable.
  * Gyms and Elite 4 can be given a type theme when chosing Random for trainer pokemon
  * A new `Randomize Settings` button has been added. This will randomly select settings available for the game (randomize the randomizer). You will still be required to click `Randomize (Save)` to confirm the settings and generate the ROM.
+ * Fix Minimum Catch Rate checkbox to be selectable regardless of what the Wild Pokemon radio button is set to